action/index was renamed to action/books
