+++ client/src/components/ui/book/containers/VisibleBookForm.js
@@ -1,7 +1,7 @@
 import React, { connect } from 'react-redux'
 import { BookForm } from '../BookForm'
 import { fetchAuthors } from '../../../actions/authors'
-import { updateNewOrExistingBookProperty, submitBookWithValidation } from '../../../actions/index'
+import { updateNewOrExistingBookProperty, submitBookWithValidation } from '../../../actions/books'
 
 const mapStateToProps = state => {
   console.log('state->', state)
